Attempt to fix bug resulting in intersecting parts (server crashes)
I saw a server crash with error `Part 20230603_61687_61697_1 intersects
part 20230603_61690_61695_1 (state Active) It is a bug.
(LOGICAL_ERROR)`. This was using v22.12. I can see that since then the
error message was adjusted, but I don't believe the bug itself has been
resolved.

I believe the bug comes from this change (c976b28104265dc03041bd8a3755de3f1c729794).
It's associated with this pull request: (https://github.com/ClickHouse/ClickHouse/pull/41145).
Specifically changes to the file `src/Storages/MergeTree/MergeTreeData.cpp`
seem maybe incorrect to me, since the logic used when a part is covered
by another part seems to have changed as described below.

The change modifies the function `renameTempPartAndReplaceImpl`, such
that a call to `getPartHierarchy` is used where previously there was a
call to `getActivePartsToReplace`. These two functions have a similar
implementation, but one notable difference -- in the latter, execution
breaks early if a covering part is found over the new part. In our outer
function, there was a condition looking for this covering part, which if
hit, would print the warning and return false.

Without the check, it instead continues, sees that there is an
intersecting part, and crashed with the exception. I believe that the
"intersecting" part would also be considered a "covering" part, as in
this case it covers the block range from before -> after the range of
the new part.

I believe a solution is to simply re-add the condition which checks for
a covering part, and return early in that case. However, my assessment
could be very wrong, I can't say I understand this code all that well. I
also don't really know how I could write a test for this either.

The other relevant detail; This occurred at the moment when I was
running a mutation on a *different* partition of the table. It happened
on two occasions, so I'm pretty sure connected to the mutation, although
I don't know why operations in different partitions would affect each
other in this way.